+++ src/eg/edu/guc/supermarket/products/DairyProduct.java
@@ -3,7 +3,7 @@ package eg.edu.guc.supermarket.products;
 public class DairyProduct extends GroceryProduct{
 	FatLevel f;
 	
-	public DairyProduct(String name, double price, int discount, FatLevel f){
+	public DairyProduct(String name, double price, double discount, FatLevel f){
 		super(name, price, discount);
 		this.f = f;
 	}
